added more files for the bitcount program to compare & contrast
there is a counterpart to bitcount.c in here which uses the
inefficient version of bitcount. i also added some assembly files
generated from both of the C files to show the difference between the
two algorithms. one can run a `diff' on the asm files to see exactly
what the differences are.